NodeIdGenerator: widen node ids from uint16_t to uint32_t
The id space needs to match NodeHandle (uint32_t) and comfortably clear
the ~65k concurrent-node ceiling that uint16_t imposed.

+++ leylib/NodeIdGenerator.cpp
@@ -1,8 +1,8 @@
 #include <leylib/NodeIdGenerator.h>
 
-uint16_t NodeIdGenerator::generateId() {
+uint32_t NodeIdGenerator::generateId() {
   if (!availableIds_.empty()) {
-    uint16_t id = *availableIds_.begin();
+    uint32_t id = *availableIds_.begin();
     availableIds_.erase(availableIds_.begin());
     return id;
   }
@@ -10,4 +10,4 @@ uint16_t NodeIdGenerator::generateId() {
   return highestId_++;
 }
 
-void NodeIdGenerator::freeId(uint16_t id) { availableIds_.insert(id); }
+void NodeIdGenerator::freeId(uint32_t id) { availableIds_.insert(id); }
