+++ leylib/include/leylib/NodeIdGenerator.h
@@ -11,12 +11,12 @@ public:
   ~NodeIdGenerator() = default;
 
   // Returns a unique id (a previously freed one if available, else a new one).
-  uint16_t generateId();
+  uint32_t generateId();
 
   // Returns an id to the pool so it can be reused.
-  void freeId(uint16_t id);
+  void freeId(uint32_t id);
 
 private:
-  uint16_t highestId_{0};
-  std::unordered_set<uint16_t> availableIds_{};
+  uint32_t highestId_{0};
+  std::unordered_set<uint32_t> availableIds_{};
 };
